CRUD Order/Order: controller, service, request, resource

+++ app/Http/Controllers/Order/OrderController.php
@@ -2,16 +2,19 @@
 
 namespace App\Http\Controllers\Order;
 
+use App\Http\Requests\Order\IndexOrderRequest;
 use App\Http\Requests\Order\StoreOrderRequest;
 use App\Http\Requests\Order\UpdateOrderRequest;
+use App\Http\Resources\OrderResource;
 use App\Models\Order\Order;
+use App\Services\Order\OrderService;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 
 class OrderController extends Controller
 {
-  
+
     protected OrderService $OrderService;
 
     public function __construct(OrderService $OrderService)
@@ -20,18 +23,20 @@ class OrderController extends Controller
     }
 
     /**
-     * Display a listing of the resource.
-     * @throws \Exception
+     * Display a listing of the orders.
+     * @param \App\Http\Requests\Order\IndexOrderRequest $request
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function index(Request $request): JsonResponse
+    public function index(IndexOrderRequest $request): JsonResponse
     {
-        $orders = $this->OrderService->getOrders($request);
-        return self::paginated($orders, 'Orders retrieved successfully', 200);
+        $orders = $this->OrderService->getOrders($request->validated());
+        return self::paginated($orders, OrderResource::class, 'Orders retrieved successfully', 200);
     }
 
     /**
      * Store a newly created resource in storage.
-     * @throws \Exception
+     * @param \App\Http\Requests\Order\StoreOrderRequest $request
+     * @return \Illuminate\Http\JsonResponse
      */
     public function store(StoreOrderRequest $request): JsonResponse
     {
@@ -40,49 +45,64 @@ class OrderController extends Controller
     }
 
     /**
-     * Display the specified resource.
+     * Display the specified order.
+     * @param \App\Models\Order\Order $order
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     * @return \Illuminate\Http\JsonResponse
      */
     public function show(Order $order): JsonResponse
     {
-        return self::success($order, 'Order retrieved successfully');
+        if ($order && $order->user_id !== Auth::id()) {
+            return self::error(null, 'You do not have permission to access this resource.', 403);
+        }
+        return self::success(new OrderResource($order), 'Order retrieved successfully');
     }
 
     /**
-     * Update the specified resource in storage.
-     * @throws \Exception
+     * Update the specified order in storage.
+     * @param \App\Http\Requests\Order\UpdateOrderRequest $request
+     * @param \App\Models\Order\Order $order
+     * @return \Illuminate\Http\JsonResponse
      */
     public function update(UpdateOrderRequest $request, Order $order): JsonResponse
     {
         $updatedOrder = $this->OrderService->updateOrder($order, $request->validated());
-        return self::success($updatedOrder, 'Order updated successfully');
+        return $updatedOrder['status']
+            ? self::success($updatedOrder['order'], 'Order updated successfully')
+            : self::error(null, $updatedOrder['msg'], $updatedOrder['code']);
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove the specified order from storage.
+     * @param \App\Models\Order\Order $order
+     * @return JsonResponse
      */
-    public function destroy(Order $order): JsonResponse
+    public function destroy(Order $order)
     {
-        $order->delete();
-        return self::success(null, 'Order deleted successfully');
+        $destroiedOrder = $this->OrderService->destroyOrder($order);
+        return $destroiedOrder['status']
+            ? self::success(null, 'Order deleted successfully')
+            : self::error(null, $destroiedOrder['msg'], $destroiedOrder['code']);
     }
 
     /**
      * Display soft-deleted records.
+     * @return \Illuminate\Http\JsonResponse
      */
     public function showDeleted(): JsonResponse
     {
-        $orders = Order::onlyTrashed()->get();
-        return self::success($orders, 'Orders retrieved successfully');
+        $orders = Order::onlyTrashed()->where('user_id', Auth::id())->get();
+        return self::success(OrderResource::collection($orders), 'Orders retrieved successfully');
     }
 
     /**
      * Restore a soft-deleted record.
      * @param string $id
-     * @return JsonResponse
+     * @return \Illuminate\Http\JsonResponse
      */
     public function restoreDeleted(string $id): JsonResponse
     {
-        $order = Order::onlyTrashed()->findOrFail($id);
+        $order = Order::onlyTrashed()->where('user_id', Auth::id())->findOrFail($id);
         $order->restore();
         return self::success($order, 'Order restored successfully');
     }
@@ -90,11 +110,11 @@ class OrderController extends Controller
     /**
      * Permanently delete a soft-deleted record.
      * @param string $id
-     * @return JsonResponse
+     * @return \Illuminate\Http\JsonResponse
      */
     public function forceDeleted(string $id): JsonResponse
     {
-        $order = Order::onlyTrashed()->findOrFail($id)->forceDelete();
+        $order = Order::onlyTrashed()->where('user_id', Auth::id())->findOrFail($id)->forceDelete();
         return self::success(null, 'Order force deleted successfully');
     }
 }
